[Refactoring] Extracted Shared logic from strategy pattern
